Move power transforms into nested module
