+++ crates/augurs-forecaster/src/transforms.rs
@@ -2,13 +2,18 @@
 Data transformations.
 */
 
+mod power;
+
 use argmin::core::Error;
 use augurs_core::{
     interpolate::{InterpolateExt, LinearInterpolator},
     Forecast,
 };
 
-use crate::power_transforms::{optimize_box_cox_lambda, optimize_yeo_johnson_lambda};
+use power::{
+    optimize_box_cox_lambda, optimize_yeo_johnson_lambda, BoxCoxExt, InverseBoxCoxExt,
+    InverseYeoJohnsonExt, YeoJohnsonExt,
+};
 
 /// Transforms and Transform implementations.
 ///
@@ -614,203 +619,6 @@ trait ExpExt: Iterator<Item = f64> {
 
 impl<T> ExpExt for T where T: Iterator<Item = f64> {}
 
-/// Returns the Box-Cox transformation of the given value.
-/// Assumes x > 0.
-pub fn box_cox(x: f64, lambda: f64) -> Result<f64, &'static str> {
-    if x <= 0.0 {
-        return Err("x must be greater than 0");
-    }
-    if lambda == 0.0 {
-        Ok(x.ln())
-    } else {
-        Ok((x.powf(lambda) - 1.0) / lambda)
-    }
-}
-
-/// Returns the Yeo-Johnson transformation of the given value.
-pub fn yeo_johnson(x: f64, lambda: f64) -> Result<f64, &'static str> {
-    if x.is_nan() || lambda.is_nan() {
-        return Err("Input values must be valid numbers.");
-    }
-
-    if x >= 0.0 {
-        if lambda == 0.0 {
-            Ok((x + 1.0).ln())
-        } else {
-            Ok(((x + 1.0).powf(lambda) - 1.0) / lambda)
-        }
-    } else if lambda == 2.0 {
-        Ok(-(-x + 1.0).ln())
-    } else {
-        Ok(-((-x + 1.0).powf(2.0 - lambda) - 1.0) / (2.0 - lambda))
-    }
-}
-
-/// An iterator adapter that applies the Box-Cox transformation to each item.
-#[derive(Clone, Debug)]
-struct BoxCox<T> {
-    inner: T,
-    lambda: f64,
-}
-
-impl<T> Iterator for BoxCox<T>
-where
-    T: Iterator<Item = f64>,
-{
-    type Item = f64;
-    fn next(&mut self) -> Option<Self::Item> {
-        self.inner
-            .next()
-            .map(|x| box_cox(x, self.lambda).unwrap_or(f64::NAN))
-    }
-}
-
-trait BoxCoxExt: Iterator<Item = f64> {
-    fn box_cox(self, lambda: f64) -> BoxCox<Self>
-    where
-        Self: Sized,
-    {
-        BoxCox {
-            inner: self,
-            lambda,
-        }
-    }
-}
-
-impl<T> BoxCoxExt for T where T: Iterator<Item = f64> {}
-
-/// Returns the inverse Box-Cox transformation of the given value.
-fn inverse_box_cox(y: f64, lambda: f64) -> Result<f64, &'static str> {
-    if lambda == 0.0 {
-        Ok(y.exp())
-    } else {
-        let value = y * lambda + 1.0;
-        if value <= 0.0 {
-            Err("Invalid domain for inverse Box-Cox transformation")
-        } else {
-            Ok(value.powf(1.0 / lambda))
-        }
-    }
-}
-
-/// An iterator adapter that applies the inverse Box-Cox transformation to each item.
-#[derive(Clone, Debug)]
-struct InverseBoxCox<T> {
-    inner: T,
-    lambda: f64,
-}
-
-impl<T> Iterator for InverseBoxCox<T>
-where
-    T: Iterator<Item = f64>,
-{
-    type Item = f64;
-    fn next(&mut self) -> Option<Self::Item> {
-        self.inner
-            .next()
-            .map(|y| inverse_box_cox(y, self.lambda).unwrap_or(f64::NAN))
-    }
-}
-
-trait InverseBoxCoxExt: Iterator<Item = f64> {
-    fn inverse_box_cox(self, lambda: f64) -> InverseBoxCox<Self>
-    where
-        Self: Sized,
-    {
-        InverseBoxCox {
-            inner: self,
-            lambda,
-        }
-    }
-}
-
-impl<T> InverseBoxCoxExt for T where T: Iterator<Item = f64> {}
-
-/// An iterator adapter that applies the Yeo-Johnson transformation to each item.
-#[derive(Clone, Debug)]
-struct YeoJohnson<T> {
-    inner: T,
-    lambda: f64,
-}
-
-impl<T> Iterator for YeoJohnson<T>
-where
-    T: Iterator<Item = f64>,
-{
-    type Item = f64;
-    fn next(&mut self) -> Option<Self::Item> {
-        self.inner
-            .next()
-            .map(|x| yeo_johnson(x, self.lambda).unwrap_or(f64::NAN))
-    }
-}
-
-trait YeoJohnsonExt: Iterator<Item = f64> {
-    fn yeo_johnson(self, lambda: f64) -> YeoJohnson<Self>
-    where
-        Self: Sized,
-    {
-        YeoJohnson {
-            inner: self,
-            lambda,
-        }
-    }
-}
-
-impl<T> YeoJohnsonExt for T where T: Iterator<Item = f64> {}
-
-/// Returns the inverse Yeo-Johnson transformation of the given value.
-fn inverse_yeo_johnson(y: f64, lambda: f64) -> f64 {
-    const EPSILON: f64 = 1e-6;
-
-    if y >= 0.0 && lambda.abs() < EPSILON {
-        // For lambda close to 0 (positive values)
-        (y.exp()) - 1.0
-    } else if y >= 0.0 {
-        // For positive values (lambda not close to 0)
-        (y * lambda + 1.0).powf(1.0 / lambda) - 1.0
-    } else if (lambda - 2.0).abs() < EPSILON {
-        // For lambda close to 2 (negative values)
-        -(-y.exp() - 1.0)
-    } else {
-        // For negative values (lambda not close to 2)
-        -((-((2.0 - lambda) * y) + 1.0).powf(1.0 / (2.0 - lambda)) - 1.0)
-    }
-}
-
-/// An iterator adapter that applies the inverse Yeo-Johnson transformation to each item.
-#[derive(Clone, Debug)]
-struct InverseYeoJohnson<T> {
-    inner: T,
-    lambda: f64,
-}
-
-impl<T> Iterator for InverseYeoJohnson<T>
-where
-    T: Iterator<Item = f64>,
-{
-    type Item = f64;
-    fn next(&mut self) -> Option<Self::Item> {
-        self.inner
-            .next()
-            .map(|y| inverse_yeo_johnson(y, self.lambda))
-    }
-}
-
-trait InverseYeoJohnsonExt: Iterator<Item = f64> {
-    fn inverse_yeo_johnson(self, lambda: f64) -> InverseYeoJohnson<Self>
-    where
-        Self: Sized,
-    {
-        InverseYeoJohnson {
-            inner: self,
-            lambda,
-        }
-    }
-}
-
-impl<T> InverseYeoJohnsonExt for T where T: Iterator<Item = f64> {}
-
 #[cfg(test)]
 mod test {
     use augurs_testing::{assert_all_close, assert_approx_eq};
@@ -986,41 +794,5 @@ mod test {
         assert_approx_eq!(params.data_max, 3.0);
         assert_approx_eq!(params.scaled_min, 0.0);
         assert_approx_eq!(params.scaled_max, 1.0);
-    }
-
-    #[test]
-    fn box_cox_test() {
-        let data = vec![1.0, 2.0, 3.0];
-        let lambda = 0.5;
-        let expected = vec![0.0, 0.8284271247461903, 1.4641016151377544];
-        let actual: Vec<_> = data.into_iter().box_cox(lambda).collect();
-        assert_all_close(&expected, &actual);
-    }
-
-    #[test]
-    fn inverse_box_cox_test() {
-        let data = vec![0.0, 0.5_f64.ln(), 1.0_f64.ln()];
-        let lambda = 0.5;
-        let expected = vec![1.0, 0.426966072919605, 1.0];
-        let actual: Vec<_> = data.into_iter().inverse_box_cox(lambda).collect();
-        assert_all_close(&expected, &actual);
-    }
-
-    #[test]
-    fn yeo_johnson_test() {
-        let data = vec![-1.0, 0.0, 1.0];
-        let lambda = 0.5;
-        let expected = vec![-1.2189514164974602, 0.0, 0.8284271247461903];
-        let actual: Vec<_> = data.into_iter().yeo_johnson(lambda).collect();
-        assert_all_close(&expected, &actual);
-    }
-
-    #[test]
-    fn inverse_yeo_johnson_test() {
-        let data = vec![-1.2189514164974602, 0.0, 0.8284271247461903];
-        let lambda = 0.5;
-        let expected = vec![-1.0, 0.0, 1.0];
-        let actual: Vec<_> = data.into_iter().inverse_yeo_johnson(lambda).collect();
-        assert_all_close(&expected, &actual);
     }
 }
